### 2025-12-20 #9672 @editorialbot
[REVIEW]: nf-core/cutandrun: A Nextflow pipeline for the analysis of CUT&RUN, CUT&Tag and TIP-seq datasets
**Submitting author:** @tamara-hodgetts (<a href="http://orcid.org/0000-0002-3991-1683">Tamara L Hodgetts</a>)
**Repository:** https://github.com/nf-core/cutandrun
**Branch with paper.md** (empty if default branch): manuscript
**Version:** 3.2.2
