### 2025-12-23 #9672 @johnvusich
Dear @tamara-hodgetts,

The documentation for this pipeline states that the pipeline structure and downstream analysis are adapted from the original CUT&Tag analysis [pipeline](https://dx.doi.org/10.17504/protocols.io.bjk2kkye) from the Henikoff lab. Please cite the original pipeline from protocols.io in the manuscript.

[dx.doi.org/10.17504/protocols.io.bjk2kkye](https://dx.doi.org/10.17504/protocols.io.bjk2kkye)
