### 2025-12-30 0fa4974
Add TODO list for extension specifications
