### 2025-12-30 da09db3
Create TODO_PeakCallerExtension.md
