+++ TODO_PeakCallerExtension.md
@@ -1,712 +1,0 @@
-# Peak-Calling Extension Module Specification
-
-## Overview
-
-This specification defines a multi-caller peak-calling extension module for CUT&RUN data analysis. The module runs **7 peak-calling variants** across 3 tools (MACS2, epic2, SPAN/omnipeaks) on all samples, generates consensus peaks, computes QC metrics, and produces DiffBind-ready sample sheets.
-
----
-
-## Table of Contents
-
-1. [Peak-Calling Variants](#1-peak-calling-variants)
-2. [Input Requirements](#2-input-requirements)
-3. [Experimental Design](#3-experimental-design)
-4. [Genome Handling](#4-genome-handling)
-5. [Control/IgG Handling](#5-controligg-handling)
-6. [Replicate Handling](#6-replicate-handling)
-7. [Execution Model](#7-execution-model)
-8. [Output Structure](#8-output-structure)
-9. [QC Metrics and Reporting](#9-qc-metrics-and-reporting)
-10. [DiffBind Integration](#10-diffbind-integration)
-11. [Resource Requirements](#11-resource-requirements)
-12. [Error Handling](#12-error-handling)
-13. [Special Considerations](#13-special-considerations)
-
----
-
-## 1. Peak-Calling Variants
-
-### 1.1 MACS2 (2 variants) - No Control
-
-CUT&RUN-optimized peak calling without IgG background subtraction.
-
-| Variant | Command |
-|---------|---------|
-| `macs2_narrow` | `macs2 callpeak --nomodel --shift -75 --extsize 150 --keep-dup all -q 0.05 --gsize $GSIZE --format BAMPE --name $PREFIX --treatment $TREATMENT_BAM` |
-| `macs2_broad` | `macs2 callpeak --nomodel --shift -75 --extsize 150 --keep-dup all -q 0.05 --broad --broad-cutoff 0.1 --gsize $GSIZE --format BAMPE --name $PREFIX --treatment $TREATMENT_BAM` |
-
-**Key Parameters:**
-- `--nomodel --shift -75 --extsize 150`: Fixed 150bp fragment handling (CUT&RUN optimized)
-- `--keep-dup all`: Relies on upstream deduplication
-- `-q 0.05`: FDR threshold (same for narrow and broad)
-- `--broad-cutoff 0.1`: Linking threshold for broad peaks
-- **No `--control`**: Peaks called without IgG background
-
-**Environment:**
-```bash
-# Run in subshell to isolate Python 2.7 environment
-(
-  module purge
-  module load viz
-  module load biology
-  module load py-macs2/2.1.1_py27
-  module load samtools/1.8
-  # ... run macs2 command ...
-)
-```
-
-**Outputs:**
-- Narrow: `{PREFIX}_peaks.narrowPeak`
-- Broad: `{PREFIX}_peaks.broadPeak`
-- Intermediates: `{PREFIX}_model.r`, `{PREFIX}_peaks.xls`, `{PREFIX}_summits.bed`
-
----
-
-### 1.2 epic2 (3 variants) - With Control
-
-SICER2-replacement for broad domain calling with IgG background subtraction.
-
-| Variant | Window | Gaps | Command |
-|---------|--------|------|---------|
-| `epic2_200bp` | 200 | 3 (=600bp) | `epic2 --treatment $TREATMENT_BAM --control $POOLED_CONTROL_BAM --genome $GENOME -w 200 -g 3 -fdr 0.01 --output $PREFIX_epic2_200bp.peaks` |
[…]
