### 2025-12-30 78f0669
Revise ChIPBinner documentation for clarity
ChIPBinner documentation updated to enhance clarity and structure, including sections on algorithmic differences, use cases, normalization options, and workflow extensions.
