### 2025-12-30 a2de1bc
Add specification for condition-aware peak calling extension
This spec covers TODOs 1-3 from the extension roadmap:

1. Samplesheet extension with 'condition' column for experimental design
   (WT/KO, Treatment/Control) with backward compatibility

2. Group-specific handling for normalization and peak calling:
   - Condition-aware spike-in normalization
   - Control pairing respecting conditions
   - Extended consensus peak modes (by condition, group×condition)

3. Support for additional peak callers:
   - GoPeaks (narrow + broad variants)
   - epic2 (200bp, 150bp, 25bp window variants)
   - S
[…]
