+++ docs/extension_specs/SPEC_ConditionAwarePeakCalling.md
@@ -1,0 +1,842 @@
+# SPEC: Condition-Aware Peak Calling Extension
+
+## Overview
+
+This specification defines extensions to nf-core/cutandrun that add:
+1. **Condition metadata** in the samplesheet for experimental design (Treatment vs. Control)
+2. **Group-specific handling** for normalization and peak calling
+3. **Support for additional peak callers**: GoPeaks, epic2, and SPAN/omnipeaks
+
+These features enable differential analysis workflows and provide users with flexibility in choosing peak-calling strategies optimized for different histone mark profiles.
+
+---
+
+## Table of Contents
+
+1. [Motivation](#1-motivation)
+2. [Samplesheet Extension](#2-samplesheet-extension)
+3. [Group-Specific Processing](#3-group-specific-processing)
+4. [Additional Peak Callers](#4-additional-peak-callers)
+5. [Implementation Plan](#5-implementation-plan)
+6. [Configuration Parameters](#6-configuration-parameters)
+7. [Output Structure](#7-output-structure)
+8. [Testing Strategy](#8-testing-strategy)
+9. [Migration Guide](#9-migration-guide)
+
+---
+
+## 1. Motivation
+
+### 1.1 Current Limitations
+
+The current nf-core/cutandrun pipeline:
+- Treats all non-control samples as equivalent targets
+- Applies uniform normalization across all samples
+- Supports only SEACR and MACS2 for peak calling
+- Cannot distinguish between experimental conditions (e.g., WT vs. KO)
+
+### 1.2 Use Cases Enabled
+
+This extension supports:
+- **Differential analysis preparation**: Generate DiffBind-ready samplesheets with condition metadata
+- **Condition-aware normalization**: Apply appropriate scaling factors per experimental group
+- **Peak caller comparison**: Run multiple peak callers to identify optimal methods for specific marks
+- **Broad histone mark analysis**: epic2 and SPAN are better suited for broad domains (H3K36me2/me3, H3K27me3)
+
+---
+
+## 2. Samplesheet Extension
+
+### 2.1 New Samplesheet Format
+
+**Current format:**
+```csv
+group,replicate,fastq_1,fastq_2,control
+```
+
+**Extended format:**
+```csv
+group,replicate,fastq_1,fastq_2,control,condition
+```
+
+### 2.2 Column Definitions
+
+| Column | Required | Description | Example Values |
+|--------|----------|-------------|----------------|
+| `group` | Yes | Sample group identifier (typically histone mark) | `H3K36me2`, `H3K4me3` |
+| `replicate` | Yes | Replicate number (integer ≥ 1) | `1`, `2` |
+| `fastq_1` | Yes | Path to read 1 FASTQ file | `/path/to/sample_R1.fq.gz` |
+| `fastq_2` | Yes | Path to read 2 FASTQ file | `/path/to/sample_R2.fq.gz` |
+| `control` | Conditional | IgG control group name (empty for controls) | `IgG_WT`, `IgG_KO` |
+| `condition` | No* | Experimental condition | `WT`, `KO`, `Treatment`, `Control` |
+
+*`condition` is optional for backward compatibility but required for condition-aware processing.
+
+### 2.3 Example Samplesheet
+
+```csv
+group,replicate,fastq_1,fastq_2,control,condition
[…]
