### 2025-12-30 8ef9668
Implement condition-aware peak calling extension
This implements the specification from SPEC_ConditionAwarePeakCalling.md:

## Samplesheet Extension (TODO 1)
- Add optional 'condition' column to samplesheet format (backward compatible)
- Update check_samplesheet.py to parse and validate condition field
- Update input_check.nf to propagate meta.condition through pipeline

## Group-Specific Processing (TODO 2)
- Add normalisation_scope parameter (global, condition, group, group_condition)
- Add condition_aware_norm parameter for differential analysis workflows
- Add pool_controls_by parameter fo
[…]
