+++ bin/check_samplesheet.py
@@ -67,17 +67,21 @@ def check_samplesheet(file_in, file_out, use_control):
     """
     This function checks that the samplesheet follows the following structure:
 
-    group,replicate,fastq_1,fastq_2,control
-    WT,1,WT_LIB1_REP1_1.fastq.gz,WT_LIB1_REP1_2.fastq.gz,CONTROL_GROUP
-    WT,1,WT_LIB2_REP1_1.fastq.gz,WT_LIB2_REP1_2.fastq.gz,CONTROL_GROUP
-    WT,2,WT_LIB1_REP2_1.fastq.gz,WT_LIB1_REP2_2.fastq.gz,CONTROL_GROUP
-    KO,1,KO_LIB1_REP1_1.fastq.gz,KO_LIB1_REP1_2.fastq.gz,CONTROL_GROUP
-    CONTROL_GROUP,1,KO_LIB1_REP1_1.fastq.gz,IGG_LIB1_REP1_2.fastq.gz,
-    CONTROL_GROUP,2,KO_LIB1_REP1_1.fastq.gz,IGG_LIB1_REP1_2.fastq.gz,
+    group,replicate,fastq_1,fastq_2,control[,condition]
+    WT,1,WT_LIB1_REP1_1.fastq.gz,WT_LIB1_REP1_2.fastq.gz,CONTROL_GROUP,WT
+    WT,1,WT_LIB2_REP1_1.fastq.gz,WT_LIB2_REP1_2.fastq.gz,CONTROL_GROUP,WT
+    WT,2,WT_LIB1_REP2_1.fastq.gz,WT_LIB1_REP2_2.fastq.gz,CONTROL_GROUP,WT
+    KO,1,KO_LIB1_REP1_1.fastq.gz,KO_LIB1_REP1_2.fastq.gz,CONTROL_GROUP,KO
+    CONTROL_GROUP,1,KO_LIB1_REP1_1.fastq.gz,IGG_LIB1_REP1_2.fastq.gz,,WT
+    CONTROL_GROUP,2,KO_LIB1_REP1_1.fastq.gz,IGG_LIB1_REP1_2.fastq.gz,,KO
+
+    The 'condition' column is optional for backward compatibility.
+    If not provided, defaults to 'default' for all samples.
     """
 
     # Init
     control_present = False
+    condition_present = False
     num_fastq_list = []
     sample_names_list = []
     control_names_list = []
@@ -87,21 +91,31 @@ def check_samplesheet(file_in, file_out, use_control):
         ## Check header
         MIN_COLS = 3
         LEGACY_HEADER = ["group", "replicate", "control_group", "fastq_1", "fastq_2"]
-        HEADER = ["group", "replicate", "fastq_1", "fastq_2", "control"]
-        HEADER_LEN = len(HEADER)
+        HEADER_BASE = ["group", "replicate", "fastq_1", "fastq_2", "control"]
+        HEADER_EXTENDED = ["group", "replicate", "fastq_1", "fastq_2", "control", "condition"]
         header = [x.strip('"') for x in fin.readline().strip().split(",")]
 
         if len(header) >= len(LEGACY_HEADER) and header[: len(LEGACY_HEADER)] == LEGACY_HEADER:
             print(
                 "ERROR: It looks like you are using a legacy header format with a newer version of the pipeline -> {} != {}".format(
-                    ",".join(header), ",".join(HEADER)
+                    ",".join(header), ",".join(HEADER_BASE)
                 )
             )
             sys.exit(1)
 
-        if header[: len(HEADER)] != HEADER:
-            print("ERROR: Please check samplesheet header -> {} != {}".format(",".join(header), ",".join(HEADER)))
+        # Check for extended header with condition column
+        if len(header) == len(HEADER_EXTENDED) and header == HEADER_EXTENDED:
[…]
