### 2026-01-19 aa821b2
Add manuscript
