+++ paper/paper.md
@@ -1,0 +1,81 @@
+---
+title: 'nf-core/cutandrun: A Nextflow pipeline for the analysis of CUT&RUN, CUT&Tag and TIP-seq datasets'
+tags:
+  - Nextflow
+  - nf-core
+  - CUT&RUN
+  - CUT&Tag
+  - TIP-seq
+  - single-cell
+  - epigenomics
+authors:
+  - name: Tamara L. Hodgetts
+    orcid: 0000-0002-3991-1683
+    corresponding: true
+    affiliation: 1 
+  - name: Charlotte West
+    orcid: 0000-0003-2094-3324
+    affiliation: 2
+  - name: Nicholas M. Luscombe
+    orcid: 0000-0001-5293-4778
+    affiliation: 3
+  - name: Jernej Ule
+    orcid: 0000-0002-2452-4277
+    affiliation: 1
+  - name: James Briscoe
+    orcid: 0000-0002-1020-5240
+    affiliation: 1
+  - name: Chris Cheshire
+    orcid: 0000-0002-5806-2883
+    affiliation: 1
+
+affiliations:
+ - name: The Francis Crick Institute, London, UK
+   index: 1
+ - name: European Molecular Biology Laboratory, European Bioinformatics Institute (EMBL-EBI), Wellcome Genome Campus, Hinxton, UK
+   index: 2
+ - name: Okinawa Institute of Science and Technology, Okinawa, Japan
+   index: 3
+date: 7 March 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+Mapping transcription factor binding events and histone modifications on a genome-wide scale is key to understanding the regulatory dynamics controlling gene expression. Recently, CUT&RUN, CUT&Tag and TIP-seq protocols provide new avenues to map chromatin-associated proteins, offering advantages compared to ChIP-seq in terms of sensitivity, resolution and sample consumption. Despite differences in biochemistry, the core computational analysis of CUT&RUN, CUT&Tag and TIP-seq data exhibit notable similarities, representing an opportunity for a unified bioinformatics solution. 
+
+Here, we present nf-core/cutandrun, a best-practice bioinformatics analysis workflow for CUT&RUN, CUT&Tag and TIP-seq data. In contrast to existing alternative pipelines, nf-core/cutandrun was developed based on the nf-core community framework to provide enhanced reproducibility, flexibility, scalability, portability and robustness. nf-core/cutandrun additionally enables the specification of spike-in genomes with different normalisation options, supports multiple peak callers, and provides extensive quality control metrics reporting. The pipeline supports a wide range of execution environments and operating systems, and is usable with minimal technical knowledge. We encourage user engagement with the wider nf-core community through the official nf-core GitHub repository and Slack channels, and aim to incorporate feedback in subsequent releases to ensure that the pipeline remains dynamic and up-to-date.
+
+
+# Statement of need
+
[…]
