+++ paper/paper.bib
@@ -1,0 +1,203 @@
+@article{Skene:2017,
+  	Author = {{Skene}, PJ. and {Henikoff}, S.},
+  	Title = {{An efficient targeted nuclease strategy for high-resolution mapping of DNA binding sites}},
+    Journal = {eLife},
+  	Year = {2017},
+    Volume = {6},
+    Pages = {e21856},
+	doi = {10.7554/eLife.21856}
+}
+
+@article{Bartlett:2021,
+  	Author = {{Bartlett}, DA. and {Dileep}, V. and {Handa}, T. and {Ohkawa}, Y. and {Kimura}, H. and {Henikoff}, S. and {Gilbert}, DM.},
+  	Title = {{High-throughput single-cell epigenomic profiling by targeted insertion of promoters (TIP-seq)}},
+    Journal = {Journal of Cell Biology},
+  	Year = {2021},
+    Volume = {220},
+    Pages = {e202103078},
+	doi = {10.1083/jcb.202103078}
+}
+
+@article{Daunesse:2022,
+  	Author = {{Daunesse}, M. and {Legendre}, R. and {Varet}, H. and {Pain}, A. and {Chica}, C.},
+  	Title = {{ePeak: from replicated chromatin profiling data to epigenomic dynamics}},
+    Journal = {NAR Genomics and Bioinformatics},
+  	Year = {2022},
+    Volume = {4},
+    Pages = {lqac041},
+	doi = {10.1093/nargab/lqac041}
+}
+
+@article{Tommaso:2017,
+  	Author = {{Di Tommaso}, P. and {Chatzou}, M. and {Floden}, EW. and {Barja}, PP. and {Palumbo}, E. and {Notredame}, C.},
+  	Title = {{Nextflow enables reproducible computational workflows}},
+    Journal = {Nature Biotechnology},
+  	Year = {2017},
+    Volume = {35},
+    Pages = {316-9},
+	doi = {10.1038/nbt.3820}
+}
+
+@article{Ewels:2016,
+  	Author = {{Ewels}, PA. and {Magnusson}, M. and {Lundin}, S. and {Käller}, M.},
+  	Title = {{MultiQC: summarize analysis results for multiple tools and samples in a single report}},
+    Journal = {Bioinformatics},
+  	Year = {2016},
+    Volume = {32},
+    Pages = {3047-8},
+	doi = {10.1093/bioinformatics/btw354}
+}
+
+@article{Ewels:2020,
+  	Author = {{Ewels}, PA. and {Peltzer}, A. and {Fillinger}, S. and {Patel}, H. and {Alneberg}, J. and {Wilm}, A. and {Garcia}, MU. and {Di Tommaso}, P. and {Nahnsen}, S.},
+  	Title = {{The nf-core framework for community-curated bioinformatics pipelines}},
+    Journal = {Nature Biotechnology},
+  	Year = {2020},
+    Volume = {38},
+    Pages = {276-8},
+	doi = {10.1038/s41587-020-0439-x}
+}
+
+@article{Furey:2012,
+  	Author = {{Furey}, TS.},
+  	Title = {{ChIP–seq and beyond: new and improved methodologies to detect and characterize protein–DNA interactions.}},
+    Journal = {Nature Reviews Genetics},
+  	Year = {2012},
+    Volume = {13},
+    Pages = {840-52},
+	doi = {10.1038/nrg3306}
+}
+
+@article{Kaya-Okur:2019,
+  	Author = {{Kaya-Okur}, HS. and {Wu}, SJ. and {Codomo}, CA. and {Pledger}, ES. and {Bryson}, TD. and {Henikoff}, JG. and {Ahmad}, K. and {Henikoff}, S.},
[…]
