### 2026-01-19 443bd68
Added protocol ref and reduced word count
