+++ paper/paper.md
@@ -44,14 +44,14 @@ bibliography: paper.bib
 
 Mapping transcription factor binding events and histone modifications on a genome-wide scale is key to understanding the regulatory dynamics controlling gene expression. Recently, CUT&RUN, CUT&Tag and TIP-seq protocols provide new avenues to map chromatin-associated proteins, offering advantages compared to ChIP-seq in terms of sensitivity, resolution and sample consumption. Despite differences in biochemistry, the core computational analysis of CUT&RUN, CUT&Tag and TIP-seq data exhibit notable similarities, representing an opportunity for a unified bioinformatics solution. 
 
-Here, we present nf-core/cutandrun, a best-practice bioinformatics analysis workflow for CUT&RUN, CUT&Tag and TIP-seq data. In contrast to existing alternative pipelines, nf-core/cutandrun was developed based on the nf-core community framework to provide enhanced reproducibility, flexibility, scalability, portability and robustness. nf-core/cutandrun additionally enables the specification of spike-in genomes with different normalisation options, supports multiple peak callers, and provides extensive quality control metrics reporting. The pipeline supports a wide range of execution environments and operating systems, and is usable with minimal technical knowledge. We encourage user engagement with the wider nf-core community through the official nf-core GitHub repository and Slack channels, and aim to incorporate feedback in subsequent releases to ensure that the pipeline remains dynamic and up-to-date.
+Here, we present nf-core/cutandrun, a best-practice bioinformatics pipeline for CUT&RUN, CUT&Tag and TIP-seq data. In contrast to existing alternative pipelines, nf-core/cutandrun was developed based on the nf-core community framework to provide enhanced reproducibility, flexibility, scalability, portability and robustness. nf-core/cutandrun additionally enables the specification of spike-in genomes with different normalisation options, supports multiple peak callers, and provides extensive quality control metrics reporting. The pipeline supports a wide range of execution environments and operating systems, and is usable with minimal technical knowledge. We encourage user engagement with the wider nf-core community through the official nf-core GitHub repository and Slack channels, and aim to incorporate feedback in subsequent releases to ensure that the pipeline remains dynamic and up-to-date.
 
 
 # Statement of need
 
 Characterising the genome-wide distribution of transcription factor binding events and histone modifications is key to elucidating the mechanics and dynamics of gene expression regulation. In recent years, the CUT&RUN and CUT&Tag protocols have emerged as alternatives to chromatin immunoprecipitation (ChIP), providing enhanced signal-to-noise ratios, as well as reduced costs and input sample requirements [@Skene:2017; @Kaya-Okur:2019; see for a review @Furey:2012; @Park:2009]. The CUT&RUN and CUT&Tag protocols first expose permeabilised, unfixed cells to primary antibodies, which bind a particular DNA-associated transcription factor or histone modification of interest. Secondary antibodies are added for signal amplification, and to facilitate the binding of either Protein A-MNase (CUT&RUN) or Protein A-Tn5 transposase (CUT&Tag) fusion constructs. Enzymatic activation subsequently enables local DNA cleavage to generate protein-associated chromatin fragments, which is followed by DNA extraction, library preparation, amplification and sequencing. Most recently, the TIP-seq protocol was additionally developed, which combines CUT&Tag with *in vitro* transcription to provide a further 10-fold increase in sensitivity [@Bartlett:2021]. 
 
[…]
