+++ paper/paper.bib
@@ -201,3 +201,11 @@
   	Url = {https://broadinstitute.github.io/picard/},
   	Note = {(Accessed: 4 March 2025)}
 }
+
+@misc{Zheng:2020,
+  	Author = {{Zheng}, Y. and {Ahmad}, K. and {Henikoff}, S.},
+  	Title = {{CUT&Tag Data Processing and Analysis Tutorial}},
+  	Year = {2020},
+  	Url = {https://www.protocols.io/view/cut-amp-tag-data-processing-and-analysis-tutorial-e6nvw93x7gmk/v1},
+  	Note = {(Accessed: 19 January 2026)}
+}
