### 2026-01-20 9abdd4a
Minor wording adjustment
