+++ paper/paper.md
@@ -51,7 +51,7 @@ Here, we present nf-core/cutandrun, a best-practice bioinformatics pipeline for 
 
 Characterising the genome-wide distribution of transcription factor binding events and histone modifications is key to elucidating the mechanics and dynamics of gene expression regulation. In recent years, the CUT&RUN and CUT&Tag protocols have emerged as alternatives to chromatin immunoprecipitation (ChIP), providing enhanced signal-to-noise ratios, as well as reduced costs and input sample requirements [@Skene:2017; @Kaya-Okur:2019; see for a review @Furey:2012; @Park:2009]. The CUT&RUN and CUT&Tag protocols first expose permeabilised, unfixed cells to primary antibodies, which bind a particular DNA-associated transcription factor or histone modification of interest. Secondary antibodies are added for signal amplification, and to facilitate the binding of either Protein A-MNase (CUT&RUN) or Protein A-Tn5 transposase (CUT&Tag) fusion constructs. Enzymatic activation subsequently enables local DNA cleavage to generate protein-associated chromatin fragments, which is followed by DNA extraction, library preparation, amplification and sequencing. Most recently, the TIP-seq protocol was additionally developed, which combines CUT&Tag with *in vitro* transcription to provide a further 10-fold increase in sensitivity [@Bartlett:2021]. 
 
-Here, we present nf-core/cutandrun, a best-practice bioinformatics pipeline for CUT&RUN, CUT&Tag and TIP-seq datasets based on the original protocol by @Zheng:2020. To meet the modern requirements for bioinformatics analysis, which we define to be reproducibility, flexibility, scalability, portability and robust reporting, we developed the pipeline in line with the nf-core community framework [@Ewels:2020; @Tommaso:2017]. nf-core/cutandrun provides several advantages relative to CUT&RUNTools and ePeak, which are the only current alternative pipelines for the end-to-end analysis of CUT&RUN and CUT&Tag data [@Zhu:2019; @Yu:2022; @Daunesse:2022]. First, as a result of the protein production process in original CUT&RUN and CUT&Tag protocols, there is residual *E. coli* DNA that can be used as a natural spike-in calibrator. This *E. coli* DNA is increasingly purified out of commercial kits, with spike-in DNA being manually added during the protocol. Manual addition of spike-in DNA can introduce considerable experimental error, leading to low alignment rates that are insufficient for normalisation. Consequently, we included additional read count-based normalisation modes and support for user-defined spike-in genomes. The pipeline additionally provides three distinct peak calling methods, including SEACR, MACS2 narrow, and MACS2 broad peak modes. We selected SEACR as the default peak caller, as it has been demonstrated to provide greater specificity and efficiency of peak calling on sparse data [@Meers:2019]. Finally, nf-core/cutandrun provides extensive quality control (QC) metrics reporting to maximise confidence prior to downstream analysis, and uniquely supports the analysis of TIP-seq data. 
[…]
