### 2026-01-20 #9672 @tamara-hodgetts
Dear @abhishektiwari, @johnvusich, @Juke34 and @danielskatz,

Thank you for these helpful and thoughtful comments. We agree that clarity for new users is critical, and we appreciate you taking the time to highlight areas for improvement.

Regarding the README, we have addressed the points raised. We have restructured it to better support first-time users by adding clearer sections covering prerequisites and installation, documentation, testing, and usage examples, along with a table of contents for easier navigation. We have also added a prominent link to the main documentation at nf-core/cutandrun directly within the README, and moved the example command to a more appropriate location. These README updates are not yet present in the main branch, but will be included in the next pipeline release, subject to the standard nf-core review process. In the meantime, the updates can be found at https://github.com/nf-core/cutandrun/tree/cc_readme_upgrades.

We also note the comment regarding attribution of the original CUT&Tag data analysis workflow. This has been addressed in the manuscript, and the original data analysis workflow from the Henikoff lab (as described on protocols.io) is now explicitly cited.

With respect to the repository and forking model, we understand and appreciate the concern. To clarify our intent: the fork under the Francis Crick Institute organisation is not meant to increase visibility or to act as a parallel "official" repository. Rather, this reflects standard nf-core development practice, where features are developed on forks and subsequently merged upstream. In our case, this fork has been the active development environment during the manuscript preparation and review period.

We agree that this setup can be confusing for users, particularly when reading a paper. To mitigate this, we have created a [manuscript branch within the official nf-core repository](https://github.com/nf-core/cutandrun/tree/manuscript). The above-mentioned manuscript changes can be found on this branch. The Francis Crick Institute fork does need to continue to exist for ongoing development work, but the nf-core repository remains the correct place for users and contributors to engage with the pipeline.

We appreciate you raising these points, as they have helped us improve both the documentation and how this distinction is communicated.

Thank you again for the constructive feedback.

Best,
Tamara
