### 2026-01-20 #9672 @ManavalanG
@abhishektiwari Hi! It appears the review checklist generated above is based on the [recent policy revisions at JOSS](https://github.com/openjournals/joss-reviews/issues/9672#issuecomment-3719059858), even though this manuscript still falls under the older policy. Could you let me know how to proceed with the checklist?
