### 2026-01-22 #9672 @ManavalanG
@abhishektiwari As per the [recent comment](https://github.com/openjournals/joss-reviews/issues/9672#issuecomment-3772338286), authors have created a `manuscript` branch in [a repo](https://github.com/nf-core/cutandrun) different from the [one currently tagged in this issue](https://github.com/openjournals/joss-reviews/issues/9672#issue-3749465318). If that's agreeable, it would be great to update the URL to the new repo so that I would be able to use editorialbot commands to create PDF of the manuscript doc.
