### 2026-01-22 #9672 @tamara-hodgetts
@ManavalanG Of course, here is the recent manuscript version in pdf format: [paper.pdf](https://github.com/user-attachments/files/24795833/paper.pdf)
