### 2026-01-22 #9672 @abhishektiwari
Hi @tamara-hodgetts for review we can only use Editorial Bot generated PDF. Was any change made to location of paper? If so we need to fix that.
