### 2026-01-23 #9672 @tamara-hodgetts
Dear @abhishektiwari,

Yes, the revised manuscript is now available on the `manuscript` branch of the official nf-core/cutandrun repository: https://github.com/nf-core/cutandrun/tree/manuscript

We changed the location of the paper to address the second comment from @Juke34. Would it be possible to update the repository link associated with the current submission to reflect this change?

Many thanks for your help!
