### 2026-01-23 #9672 @abhishektiwari
@tamara-hodgetts As requested updated repository to (https://github.com/nf-core/cutandrun) and branch (manuscript) for this submission.
