### 2026-01-30 #9672 @ManavalanG
@tamara-hodgetts  I'm done with my review :) Just have a minor question.

**Minor:**

* When ran with the recent version of nextflow `25.10.3`, nf-core/cutandrun ran into this error - `ERROR ~ Cannot invoke method optional() on null object`. This has been [documented before](https://github.com/nf-core/cutandrun/issues/301), and using the older version of nextflow `24.04.3` resolved the issue. While [`nextflow.config` does specify in the manifest](https://github.com/nf-core/cutandrun/blob/6e1125d4fee4ea7c8b70ed836bb0e92a89e3305f/nextflow.config#L334-L340) that the allowed nextflow version are `!>=23.04.0`, I did not see any warning or error when I used above-mentioned nextflow versions, both of which are latter versions than those in the manifest. Why was this the case? I used `nextflow run` to run the pipeline.
