### 2026-02-03 #9672 @abhishektiwari
Hi @danielskatz  @openjournals/bcm-eics - Based on feedback authors have updated the target repository (now points to https://github.com/nf-core/cutandrun) and meets JOSS requirements. Given, all 3 reviewers have completed their checklist I will move forward to next step.
