### 2026-02-03 #9672 @tamara-hodgetts
@ManavalanG @abhishektiwari Many thanks for your review and for raising this.

The error you are seeing is due to a parsing change in Nextflow. In Nextflow 25.10.x the emit/output channel syntax is parsed slightly differently, and modifiers such as optional must be provided in key: value form. Pipelines written for earlier syntax therefore fail during parsing under 25.10.x.

We did try to enforce an upper bound on the supported Nextflow version; however, this fails because the pipeline never reaches the version check. The error is raised during Nextflow’s syntax parsing phase, before the workflow can execute any logic (including version validation). This is effectively a breaking change in Nextflow’s parser behavior.

We are updating the pipeline to be compatible in an upcoming release. In the meantime, the workaround is to follow the documented Nextflow version requirement, as running under 25.10.x will trigger this parse-time error.
