### 2026-02-04 #9672 @abhishektiwari
@tamara-hodgetts Thanks so much. At this point could you:

- [ ] Double check Paper authors and affiliations (including ORCIDs)
- [ ] Make a tagged release of your software on Github, and list the version tag of the archived version here.
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct title (should match the paper title).
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct authors and affiliations (should match the paper authors including order).
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct version (should match the above Github tag you created).
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct ORCIDs (should match the paper authors).

Once those completed, I can then move forward with recommending acceptance of the submission.
