### 2026-02-06 #9672 @tamara-hodgetts
@abhishektiwari Many thanks for the checklist.

- **Double check Paper authors and affiliations (including ORCIDs)**: These have all been double-checked and are correct.
- **Make a tagged release of your software on Github, and list the version tag of the archived version here:** The version of the pipeline associated with this submission is **v3.2.2**, which has been tagged on GitHub.
- **Please list the DOI of the archived version here:** Version v3.2.2 is archived on Zenodo here: https://zenodo.org/records/10606804 **DOI:** 10.5281/zenodo.10606804

For the remaining archival checks (title, authorship/order, affiliations, ORCIDs, and license): the Zenodo record is generated automatically via the nf-core GitHub–Zenodo integration, and ownership of the deposit currently sits with the nf-core organisation rather than the paper authors. As a result, we are not currently able to manually edit the metadata associated with this record.

To address this, we see a few possible paths:

1. Request ownership of the existing Zenodo record so we can manually align the metadata with the paper. We have already raised a support ticket with Zenodo, but the timeline is currently unclear.
2. Manually create a new Zenodo deposit that duplicates the v3.2.2 release specifically for the purposes of this paper.
3. Proceed with the existing Zenodo record as a faithful archive of the GitHub release, noting that its metadata reflects the automated nf-core release process rather than the paper.

We would be grateful for your guidance on which of these approaches would be most appropriate for JOSS at this stage.

Many thanks again for your guidance and support during the final stages of the review.
