### 2026-02-13 #9672 @tamara-hodgetts
@abhishektiwari Many thanks for the update! We have actually now been granted edit permission of the Zenodo archive. Shall I proceed by making the requested changes to the metadata directly on this archive?
