### 2026-02-13 #9672 @abhishektiwari
Hi @tamara-hodgetts - If you can make changes to metadata directly on this archive  then go for it. Otherwise, I got confirmation from EiC that matching title is not a hard requirement, we can work with default title generated by CI.
