### 2026-02-14 #9672 @abhishektiwari
@tamara-hodgetts Are there any pending changes raised as part of review that needs to be merge and released? For example this CR still open https://github.com/nf-core/cutandrun/pull/289. We may need these to be merged, tagged as new release.
