### 2026-02-14 #9672 @abhishektiwari
@tamara-hodgetts Please also see suggestions improvements in paper below.

Main Body (Edits in paper.md):
Line 12-13: "Recently, CUT&RUN, CUT&Tag and TIP-seq protocols provide new avenues" can be "Recently, CUT&RUN, CUT&Tag and TIP-seq protocols have providednew avenues". Alternatively, you can remove "Recently,".
Line  75, 77: "Samtools" should be "SAMtools", matching the tool's official name.

References (Edits in paper.bib): Please use {} explicitly to capitalise as by default JOSS formats reference titles as sentence case.

Line 132 (Andrews, 2010): "Babraham bioinformatics-FastQC" — "bioinformatics" should be capitalized
Line 176 (Rad et al., 2017): "An introduction to docker and analysis of its performance"  should be "Docker"
Line 165 (Kurtzer et al., 2017): "PLOS One", the journal's official name is "PLOS ONE" (all caps on "ONE")
Line 177: Please check if  abbreviation IJCSNS needed here.
