### 2026-02-17 0c59c73
Addressed minor corrections in main text
