### 2026-02-17 f38c022
Addressed minor formatting and capitalisation errors in references
