+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{Skene:2017,
   	Author = {{Skene}, PJ. and {Henikoff}, S.},
-  	Title = {{An efficient targeted nuclease strategy for high-resolution mapping of DNA binding sites}},
+  	Title = {{An efficient targeted nuclease strategy for high-resolution mapping of {DNA} binding sites}},
     Journal = {eLife},
   	Year = {2017},
     Volume = {6},
@@ -10,7 +10,7 @@
 
 @article{Bartlett:2021,
   	Author = {{Bartlett}, DA. and {Dileep}, V. and {Handa}, T. and {Ohkawa}, Y. and {Kimura}, H. and {Henikoff}, S. and {Gilbert}, DM.},
-  	Title = {{High-throughput single-cell epigenomic profiling by targeted insertion of promoters (TIP-seq)}},
+  	Title = {{High-throughput single-cell epigenomic profiling by targeted insertion of promoters ({TIP-seq})}},
     Journal = {Journal of Cell Biology},
   	Year = {2021},
     Volume = {220},
@@ -20,7 +20,7 @@
 
 @article{Daunesse:2022,
   	Author = {{Daunesse}, M. and {Legendre}, R. and {Varet}, H. and {Pain}, A. and {Chica}, C.},
-  	Title = {{ePeak: from replicated chromatin profiling data to epigenomic dynamics}},
+  	Title = {{{ePeak}: from replicated chromatin profiling data to epigenomic dynamics}},
     Journal = {NAR Genomics and Bioinformatics},
   	Year = {2022},
     Volume = {4},
@@ -30,7 +30,7 @@
 
 @article{Tommaso:2017,
   	Author = {{Di Tommaso}, P. and {Chatzou}, M. and {Floden}, EW. and {Barja}, PP. and {Palumbo}, E. and {Notredame}, C.},
-  	Title = {{Nextflow enables reproducible computational workflows}},
+  	Title = {{{Nextflow} enables reproducible computational workflows}},
     Journal = {Nature Biotechnology},
   	Year = {2017},
     Volume = {35},
@@ -40,7 +40,7 @@
 
 @article{Ewels:2016,
   	Author = {{Ewels}, PA. and {Magnusson}, M. and {Lundin}, S. and {Käller}, M.},
-  	Title = {{MultiQC: summarize analysis results for multiple tools and samples in a single report}},
+  	Title = {{{MultiQC}: summarize analysis results for multiple tools and samples in a single report}},
     Journal = {Bioinformatics},
   	Year = {2016},
     Volume = {32},
@@ -60,7 +60,7 @@
 
 @article{Furey:2012,
   	Author = {{Furey}, TS.},
-  	Title = {{ChIP–seq and beyond: new and improved methodologies to detect and characterize protein–DNA interactions.}},
+  	Title = {{{ChIP–seq} and beyond: new and improved methodologies to detect and characterize protein–{DNA} interactions.}},
     Journal = {Nature Reviews Genetics},
   	Year = {2012},
     Volume = {13},
@@ -70,7 +70,7 @@
 
 @article{Kaya-Okur:2019,
   	Author = {{Kaya-Okur}, HS. and {Wu}, SJ. and {Codomo}, CA. and {Pledger}, ES. and {Bryson}, TD. and {Henikoff}, JG. and {Ahmad}, K. and {Henikoff}, S.},
-  	Title = {{CUT&Tag for efficient epigenomic profiling of small samples and single cells}},
+  	Title = {{{CUT\&Tag} for efficient epigenomic profiling of small samples and single cells}},
[…]
