### 2026-02-17 #9672 @tamara-hodgetts
@abhishektiwari Many thanks for your message!
I have just made the suggested changes to the main body and references.

Regarding the pre-review checklist:

The feedback from the reviewers was mainly focussed on modifying the text of the paper, improving the README clarity and changing the link to that of the official nf-core repository. We have made these requested changes to the manuscript branch of the official nf-core repository. We will include these changes, as well as the minor wording correction noted in the currently open [CR](https://github.com/nf-core/cutandrun/pull/289), in the next release of the pipeline. We would prefer not to create a new release for these relatively minor changes alone, as the review and release process for nf-core is often quite time-consuming and generally involves significant changes to the code or pipeline architecture.

DOI: 10.5281/zenodo.10606804
