### 2026-02-17 #9672 @abhishektiwari
@tamara-hodgetts - I checked version and current version for target repository and archive is `3.2.2`  without `v` prefix. It seems nf-core repositories use version tag without `v` prefix. Hence why I am updating version to `3.2.2` from previous `v3.2.2`. Can you confirm?
