### 2026-02-17 #9672 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
