### 2026-03-15 #9672 @Kevin-Mattheus-Moerman
As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks:

👉 Please spell out `UK` as `United Kingdom`
- [x] Copyedit pass/checked for typos
