### 2026-03-17 #9672 @tamara-hodgetts
@Kevin-Mattheus-Moerman Thank you!

I noticed that on the JOSS website, "amp;" appears twice in the paper's title after the "&" symbols: "nf-core/cutandrun: A Nextflow pipeline for the analysis of CUT&amp;RUN, CUT&amp;Tag and TIP-seq datasets" (https://joss.theoj.org/papers/10.21105/joss.09672). I believe this is likely a formatting issue—would you be able to correct it?

Many thanks, Tamara
