### 2026-03-17 #9672 @Kevin-Mattheus-Moerman
@openjournals/dev can you advise how to treat the & symbols in the title?

They currently use this:

```
title: 'nf-core/cutandrun: A Nextflow pipeline for the analysis of CUT&RUN, CUT&Tag and TIP-seq datasets'
```
